fixed show,edit,destroy

+++ app/Http/Controllers/TasksController.php
@@ -20,9 +20,14 @@ class TasksController extends Controller
                 'user' => $user,
                 'tasks' => $tasks,
             ];
+        
+            return view('welcome', $data);
         }
         
-        return view('welcome', $data);
+        else {
+            return view('welcome', $data);
+        }
+        
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -47,28 +52,41 @@ class TasksController extends Controller
             'status' => $request->status,
             'content' => $request->content,
         ]); 
-
+        
         return redirect('/');
     }
 
     // getでtasks/idにアクセスされた場合の「取得表示処理」
     public function show($id)
     {
-        $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::id();
+            $task = Task::find($id);
+            if (\Auth::id() == $task->user_id)
+            {
+                return view('tasks.show', [
+                    'task' => $task,
+                ]);
+            }
+        }
+        
+        return redirect('/');
     }
 
     // getでtasks/id/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
-        $task = Task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::id();
+            $task = Task::find($id);
+            if (\Auth::id() == $task->user_id){
+                return view('tasks.edit', [
+                    'task' => $task,
+                ]);
+            }
+        }
+        
+        return redirect('/');
     }
 
     // putまたはpatchでtasks/idにアクセスされた場合の「更新処理」
@@ -90,9 +108,14 @@ class TasksController extends Controller
     // deleteでtasks/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-
+        if (\Auth::check()) {
+            $user = \Auth::id();
+            $task = Task::find($id);
+            if (\Auth::id() == $task->user_id){
+                $task->delete();
+            }
+        }
+        
         return redirect('/');
     }
 }
